id값 Integer -> Long 변경, Home 컨트롤러 삭제

+++ src/main/java/stg/onyou/repository/FeedRepository.java
@@ -7,7 +7,7 @@ import stg.onyou.model.entity.Feed;
 import java.util.List;
 
 @Repository
-public interface FeedRepository extends JpaRepository<Feed, Integer> {
+public interface FeedRepository extends JpaRepository<Feed, Long> {
 
     List<Feed> findAllByAccess(String access);
     List<Feed> findAllByAccessOrderByCreatedDesc(String access);
