+++ src/main/java/stg/onyou/repository/OrganizationRepository.java
@@ -3,5 +3,5 @@ package stg.onyou.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import stg.onyou.model.entity.Organization;
 
-public interface OrganizationRepository extends JpaRepository<Organization, Integer> {
+public interface OrganizationRepository extends JpaRepository<Organization, Long> {
 }
